Refactor. BPP Authorization flow
