+++ local.env
@@ -2,5 +2,5 @@ SELLER_APP_PORT=3010
 SELLER_APP_ID=sample_mobility_bpp
 SELLER_APP_URL=http://localhost:3010
 GATEWAY_URL=http://localhost:1010
-privateKey=sUUXM0UOgF1WuBgbG1ILQWrqtWZTeN5XMVlMX2DsqqNNRg3uO0a/htwq5aLASya+mWtPeZPwEykw4NBT1LHosQ==
-publicKey=TUYN7jtGv4bcKuWiwEsmvplrT3mT8BMpMODQU9Sx6LE=
+PRIVATE_KEY=sUUXM0UOgF1WuBgbG1ILQWrqtWZTeN5XMVlMX2DsqqNNRg3uO0a/htwq5aLASya+mWtPeZPwEykw4NBT1LHosQ==
+PUBLIC_KEY=TUYN7jtGv4bcKuWiwEsmvplrT3mT8BMpMODQU9Sx6LE=
