+++ src/controllers/SearchController.js
@@ -1,13 +1,14 @@
 import log4js from 'log4js';
 import SearchService from '../services/SearchService';
-import GenericResponse from './GenericResponse';
+import GenericResponse from '../utilities/GenericResponse';
 import authVerifier from '../utilities/SignVerify/AuthHeaderVerifier';
+import LookUpService from '../services/LookUpService';
 
 const search = async (req, res) => {
   const logger = log4js.getLogger('SearchController');
   logger.debug(`Search called with ${JSON.stringify(req.body)}`);
 
-  const publicKey = await GenericResponse.getPublicKey('BG');
+  const publicKey = await LookUpService.getPublicKey('BG');
   authVerifier.authorize(req, publicKey).then(() => {
     logger.debug('Request Authorized Successfully.');
     SearchService.search(req.body);
