+++ src/controllers/SelectController.js
@@ -1,5 +1,5 @@
 import log4js from 'log4js';
-import GenericResponse from './GenericResponse';
+import GenericResponse from '../utilities/GenericResponse';
 
 const select = (req, res) => {
   const logger = log4js.getLogger('SelectController');
